Completed homework for Lesson 5 (currently 22 tests) Partially done homework for Lesson 6 (completed with Menu class that should execute practice classes (TriangleSidesCheck,IsEvenCheck,CompareTwoNumbers). 18 tests left to complete)

+++ src/main/java/com/kitcenter/app/classwork/lesson5/Calculator.java
@@ -20,7 +20,7 @@ public class Calculator {
         return substract;
     }
 
-    public int power(int numberOne, int numberTwo) {
+    public int power(double numberOne, int numberTwo) {
         int power=1;
         for(int iterator=1; iterator<=numberTwo; iterator++){
             power *= numberOne;
